Migrate to hardware PWM
